fix: 폴더명 단수 변경
