+++ src/apis/premieres/postEntry.ts
@@ -1,6 +1,6 @@
 import { axiosInstance } from '../axiosInstance';
 import { END_POINTS } from '../../constants/api';
-import { PaymentsFormData, PaymentsFormDataResponse } from '../../types/premieres';
+import { PaymentsFormData, PaymentsFormDataResponse } from '../../types/premiere';
 
 export interface EntryParams extends PaymentsFormData {
     premiereId: string;
