+++ src/hooks/queries/usePermieresQuery.ts
@@ -1,7 +1,7 @@
 import { useQuery, UseQueryOptions } from '@tanstack/react-query';
 import { useApiError } from './useApiError';
 import { getDetailPermieres, getPremieres } from '../../apis/premieres/getPremieres';
-import { PremieresDetailDataResponse, PremieresResponse } from '../../types/premieres';
+import { PremieresDetailDataResponse, PremieresResponse } from '../../types/premiere';
 
 export const usePermieresQuery = () => {
